Revert "add graphql-client and -endpoint for publicInfo of community"
This reverts commit 8c5086e5cdb345216cf8fe6772bb5f14f67f79f4.
