--- backend/src/federation/client/FederationClient.ts
@@ -1,13 +1,0 @@
-import { FederatedCommunity as DbFederatedCommunity } from '@entity/FederatedCommunity'
-
-export type PublicInfo = {
-  name: string
-  description: string
-  createdAt: Date
-  publicKey: string
-}
-
-export interface FederationClient {
-  requestGetPublicKey(dbCom: DbFederatedCommunity): Promise<string | undefined>
-  requestGetPublicInfo(dbCom: DbFederatedCommunity): Promise<PublicInfo | undefined>
-}
